decomp: retail-evidence passes on the rebuilt game compiler
First run of scripts/retail-evidence-passes.py on cc1 1766b1bd (host-built
with -fno-strict-aliasing): extern-width-fix raises fun_0022f5b0 53.58 ->
56.88 and fun_001fbc50 81.22 -> 81.28, struct-width-fix raises set_pal_mode
61.07 -> 61.27. NON_MATCHING bodies/headers only; INCLUDE_ASM stays
authoritative. ./verify-baseline.sh: PASS (SHA-256 e0505810...c0c9).

+++ src/assembly/textbin/fun_0022f5b0.c
@@ -9,7 +9,7 @@ INCLUDE_ASM("config/us/expected/asm/assembly/textbin/fun_0022f5b0/FUN_0022f5b0.s
 
 
 extern struct M2c_D_0013E030 D_0013E030;
-extern s32 D_0015ED6C;
+extern f32 D_0015ED6C;
 extern u8 D_001D9B60[];
 extern s32 func_001F96F8();
 extern s32 func_001F9A10();
